Rename SuccessModal to PaymentModal
